requirements.txt
argparse==1.4.0
google-api-core==1.31.1
google-cloud-dataproc==3.1.2
google-cloud-storage==1.42.0
google-cloud==0.34.0
protobuf==3.17.3